updating jpa models based on new database structure

+++ jacq-common/src/main/java/org/jacq/common/model/jpa/TblImportProperties.java
@@ -1,7 +1,17 @@
 /*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
+ * Copyright 2016 wkoller.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 package org.jacq.common.model.jpa;
 
@@ -40,9 +50,9 @@ public class TblImportProperties implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Long id;
+    private Integer id;
     @Column(name = "IDPflanze")
-    private Long iDPflanze;
+    private Integer iDPflanze;
     @Size(max = 255)
     @Column(name = "species_name")
     private String speciesName;
@@ -57,23 +67,23 @@ public class TblImportProperties implements Serializable {
     public TblImportProperties() {
     }
 
-    public TblImportProperties(Long id) {
+    public TblImportProperties(Integer id) {
         this.id = id;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public Long getIDPflanze() {
+    public Integer getIDPflanze() {
         return iDPflanze;
     }
 
-    public void setIDPflanze(Long iDPflanze) {
+    public void setIDPflanze(Integer iDPflanze) {
         this.iDPflanze = iDPflanze;
     }
